+++ Onova.Tests.Dummy/Program.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
 using Onova.Services;
-using Onova.Tests.Dummy.Internal;
 
 namespace Onova.Tests.Dummy
 {
@@ -13,9 +12,12 @@ namespace Onova.Tests.Dummy
 
     public static class Program
     {
-        private static Version Version => Assembly.GetExecutingAssembly().GetName().Version;
-        private static string AssemblyDirPath => AppDomain.CurrentDomain.BaseDirectory;
+        private static Version Version => Assembly.GetExecutingAssembly().GetName().Version!;
+
+        private static string AssemblyDirPath => AppDomain.CurrentDomain.BaseDirectory!;
+
         private static string LastRunFilePath => Path.Combine(AssemblyDirPath, $"lastrun-{Version}.txt");
+
         private static string PackagesDirPath => Path.Combine(AssemblyDirPath, "Packages");
 
         private static readonly IUpdateManager UpdateManager = new UpdateManager(
@@ -26,7 +28,7 @@ namespace Onova.Tests.Dummy
         {
             // Dump arguments to file.
             // This is only accurate enough for simple inputs.
-            File.WriteAllText(LastRunFilePath, args.JoinToString(" "));
+            File.WriteAllLines(LastRunFilePath, args);
 
             // Get command name
             var command = args.FirstOrDefault();