+++ Onova/IUpdateManager.cs
@@ -12,9 +12,14 @@ namespace Onova
     public interface IUpdateManager : IDisposable
     {
         /// <summary>
+        /// Information about the assembly, for which the updates are managed.
+        /// </summary>
+        AssemblyMetadata Updatee { get; }
+
+        /// <summary>
         /// Checks for updates.
         /// </summary>
-        Task<CheckForUpdatesResult> CheckForUpdatesAsync();
+        Task<CheckForUpdatesResult> CheckForUpdatesAsync(CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Checks whether an update to given version has been prepared.
@@ -34,7 +39,8 @@ namespace Onova
 
         /// <summary>
         /// Launches an external executable that will apply an update to given version, once this application exits.
+        /// The updater can be instructed to also restart the application after it's updated.
         /// </summary>
-        void LaunchUpdater(Version version, bool restart = true);
+        void LaunchUpdater(Version version, bool restart, string restartArguments);
     }
 }