+++ Onova/Models/AssemblyMetadata.cs
@@ -38,19 +38,29 @@ namespace Onova.Models
     {
         /// <summary>
         /// Extracts assembly metadata from given assembly.
+        /// The specified path is used to override the executable file path in case the assembly is not meant to run directly.
         /// </summary>
-        public static AssemblyMetadata FromAssembly(Assembly assembly)
+        public static AssemblyMetadata FromAssembly(Assembly assembly, string assemblyFilePath)
         {
-            var name = assembly.GetName().Name;
-            var version = assembly.GetName().Version;
-            var filePath = assembly.Location;
+            var name = assembly.GetName().Name!;
+            var version = assembly.GetName().Version!;
+            var filePath = assemblyFilePath;
 
             return new AssemblyMetadata(name, version, filePath);
         }
 
         /// <summary>
+        /// Extracts assembly metadata from given assembly.
+        /// </summary>
+        public static AssemblyMetadata FromAssembly(Assembly assembly) => FromAssembly(assembly, assembly.Location);
+
+        /// <summary>
         /// Extracts assembly metadata from entry assembly.
         /// </summary>
-        public static AssemblyMetadata FromEntryAssembly() => FromAssembly(Assembly.GetEntryAssembly());
+        public static AssemblyMetadata FromEntryAssembly()
+        {
+            var assembly = Assembly.GetEntryAssembly() ?? throw new InvalidOperationException("Can't get entry assembly.");
+            return FromAssembly(assembly);
+        }
     }
 }