+++ Onova/Services/AggregatePackageResolver.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Onova.Exceptions;
-using Onova.Internal;
+using Onova.Internal.Extensions;
 
 namespace Onova.Services
 {
@@ -32,26 +32,28 @@ namespace Onova.Services
         }
 
         /// <inheritdoc />
-        public async Task<IReadOnlyList<Version>> GetPackageVersionsAsync()
+        public async Task<IReadOnlyList<Version>> GetPackageVersionsAsync(CancellationToken cancellationToken = default)
         {
             var aggregateVersions = new HashSet<Version>();
 
             // Get unique package versions provided by all resolvers
             foreach (var resolver in _resolvers)
             {
-                var versions = await resolver.GetPackageVersionsAsync();
+                var versions = await resolver.GetPackageVersionsAsync(cancellationToken);
                 aggregateVersions.AddRange(versions);
             }
 
             return aggregateVersions.ToArray();
         }
 
-        private async Task<IPackageResolver?> GetResolverForPackageAsync(Version version)
+        private async Task<IPackageResolver?> TryGetResolverForPackageAsync(
+            Version version,
+            CancellationToken cancellationToken)
         {
             // Try to find the first resolver that has this package version
             foreach (var resolver in _resolvers)
             {
-                var versions = await resolver.GetPackageVersionsAsync();
+                var versions = await resolver.GetPackageVersionsAsync(cancellationToken);
                 if (versions.Contains(version))
                     return resolver;
             }
@@ -65,8 +67,8 @@ namespace Onova.Services
             IProgress<double>? progress = null, CancellationToken cancellationToken = default)
         {
             // Find a resolver that has this package version
-            var resolver = await GetResolverForPackageAsync(version);
-            if (resolver == null)
+            var resolver =
+                await TryGetResolverForPackageAsync(version, cancellationToken) ??
                 throw new PackageNotFoundException(version);
 
             // Download package