+++ Onova/Services/NugetPackageResolver.cs
@@ -6,9 +6,9 @@ using System.Net;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using Newtonsoft.Json.Linq;
 using Onova.Exceptions;
 using Onova.Internal;
+using Onova.Internal.Extensions;
 
 namespace Onova.Services
 {
@@ -37,51 +37,47 @@ namespace Onova.Services
         /// Initializes an instance of <see cref="NugetPackageResolver"/>.
         /// </summary>
         public NugetPackageResolver(string serviceIndexUrl, string packageId)
-            : this(HttpClientEx.GetSingleton(), serviceIndexUrl, packageId)
+            : this(Http.Client, serviceIndexUrl, packageId)
         {
         }
 
-        private async Task<string> GetPackageBaseAddressResourceUrlAsync()
+        private async Task<string> GetPackageBaseAddressResourceUrlAsync(CancellationToken cancellationToken)
         {
             // Get all available resources
-            var response = await _httpClient.GetStringAsync(_serviceIndexUrl);
-            var resourcesJson = JToken.Parse(response)["resources"];
+            var responseJson = await _httpClient.GetJsonAsync(_serviceIndexUrl, cancellationToken);
+            var resourcesJson = responseJson.GetProperty("resources");
 
             // Get URL of the required resource
-            var expectedResourceType = "PackageBaseAddress/3.0.0";
-            foreach (var resourceJson in resourcesJson)
+            foreach (var resourceJson in resourcesJson.EnumerateArray())
             {
                 // Check resource type
-                var resourceType = resourceJson["@type"].Value<string>();
-                if (resourceType == expectedResourceType)
-                    return resourceJson["@id"].Value<string>();
+                var resourceType = resourceJson.GetProperty("@type").GetString();
+                if (string.Equals(resourceType, "PackageBaseAddress/3.0.0", StringComparison.OrdinalIgnoreCase))
+                    return resourceJson.GetProperty("@id").GetString();
             }
 
             // Resource not found
-            throw new InvalidOperationException($"[{expectedResourceType}] resource not found in service index.");
+            throw new InvalidOperationException("Expected resource not found in service index.");
         }
 
         /// <inheritdoc />
-        public async Task<IReadOnlyList<Version>> GetPackageVersionsAsync()
+        public async Task<IReadOnlyList<Version>> GetPackageVersionsAsync(CancellationToken cancellationToken = default)
         {
             // Get package base address resource URL
-            var resourceUrl = await GetPackageBaseAddressResourceUrlAsync();
+            var resourceUrl = await GetPackageBaseAddressResourceUrlAsync(cancellationToken);
 
             // Get versions
             var request = $"{resourceUrl}/{PackageIdNormalized}/index.json";
-            var response = await _httpClient.GetStringAsync(request);
-            var versionsJson = JToken.Parse(response)["versions"];
+            var responseJson = await _httpClient.GetJsonAsync(request, cancellationToken);
+            var versionsJson = responseJson.GetProperty("versions");
             var versions = new HashSet<Version>();
 
-            foreach (var versionJson in versionsJson)
+            foreach (var versionJson in versionsJson.EnumerateArray())
             {
-                // Try to parse version
-                var versionText = versionJson.Value<string>();
-                if (!Version.TryParse(versionText, out var version))
-                    continue;
+                var versionText = versionJson.GetString();
 
-                // Add to list
-                versions.Add(version);
+                if (Version.TryParse(versionText, out var version))
+                    versions.Add(version);
             }
 
             return versions.ToArray();
@@ -92,7 +88,7 @@ namespace Onova.Services
             IProgress<double>? progress = null, CancellationToken cancellationToken = default)
         {
             // Get package base address resource URL
-            var resourceUrl = await GetPackageBaseAddressResourceUrlAsync();
+            var resourceUrl = await GetPackageBaseAddressResourceUrlAsync(cancellationToken);
 
             // Get package URL
             var packageUrl = $"{resourceUrl}/{PackageIdNormalized}/{version}/{PackageIdNormalized}.{version}.nupkg";
@@ -108,10 +104,8 @@ namespace Onova.Services
             response.EnsureSuccessStatusCode();
 
             // Copy content to file
-            using var input = await response.Content.ReadAsFiniteStreamAsync();
             using var output = File.Create(destFilePath);
-
-            await input.CopyToAsync(output, progress, cancellationToken);
+            await response.Content.CopyToStreamAsync(output, progress, cancellationToken);
         }
     }
 }